backend/requirements.txt
# Principais
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
passlib==1.7.4
email-validator==2.0.0.post2

# Banco de dados
sqlalchemy==2.0.20
asyncpg==0.28.0
alembic==1.12.0
psycopg2-binary==2.9.7

# Autenticação
bcrypt==4.0.1

# Validação e processamento de dados
python-dateutil==2.8.2
python-slugify==8.0.1
Pillow==10.0.0
fpdf==1.7.2
PyPDF2==3.0.1
openpyxl==3.1.2
pycep-correios==5.2.0
pandas>=2.0.0
pdfkit>=1.0.0

# Cache e limitação de taxa
redis==5.0.1
fakeredis==2.20.0

# Monitoramento e logging
prometheus-fastapi-instrumentator==6.1.0
pydantic[email]==2.3.0

# Utilitários
requests==2.31.0
httpx==0.24.1
pytz==2023.3

# Desenvolvimento
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
coverage==7.3.1
black==23.9.1
isort==5.12.0
flake8==6.1.0
bandit==1.7.5
mypy==1.5.1
pre-commit==3.4.0

# Documentação
pyyaml==6.0.1
jinja2==3.1.2

# CORS, compressão e segurança
starlette==0.27.0
aiofiles==23.2.1
cryptography==41.0.3
python-magic==0.4.27

# Health Check e Monitoramento
psutil==5.9.5
prometheus-client==0.17.1

# Segurança adicional
python-multipart==0.0.6
secure==0.3.0

# Dependências para produção
gunicorn==21.2.0
uvloop==0.18.0; sys_platform != 'win32'
httptools==0.6.0
sentry-sdk==1.28.1 